GET, POST /threeStyle を別ファイルに切り出した
