+++ index.js
@@ -12,7 +12,6 @@ const path = require('path');
 const Sequelize = require('sequelize');
 const route = require('./src/route');
 const validation = require('./src/validation');
-const utils = require('./src/lib/utils');
 const { badRequestError, } = require('./src/lib/utils');
 const { sequelize, } = require('./src/model');
 
@@ -519,106 +518,7 @@ sequelize.sync().then(() => {
             });
     });
 
-    app.get(process.env.EXPRESS_ROOT + '/threeStyle/:part', (req, res, next) => {
-        const userName = req.query.userName;
-        const buffer = req.query.buffer;
-        const sticker1 = req.query.sticker1;
-        const sticker2 = req.query.sticker2;
-        const part = req.params.part;
-        // const setup = req.query.setup;
-        // const move1 = req.query.move1;
-        // const move2 = req.query.move2;
-
-        if (!(part === 'corner' || part === 'edgeMiddle')) {
-            logger.emit('api.request', {
-                requestType: 'GET',
-                endpoint: '/hinemos/threeStyle/',
-                params: {
-                    userName,
-                    part,
-                    buffer,
-                    sticker1,
-                    sticker2,
-                },
-                status: 'error',
-                code: 400,
-                msg: '',
-            });
-
-            res.status(400).send(badRequestError);
-            return;
-        }
-
-        const query = {
-            where: {},
-        };
-        if (userName) {
-            query.where.userName = userName;
-        }
-        if (buffer) {
-            query.where.buffer = buffer;
-        }
-        if (sticker1) {
-            query.where.sticker1 = sticker1;
-        }
-        if (sticker2) {
-            query.where.sticker2 = sticker2;
-        }
-
-        let threeStyleModel;
-        if (part === 'corner') {
-            threeStyleModel = ThreeStyleCorner;
-        } else if (part === 'edgeMiddle') {
-            threeStyleModel = ThreeStyleEdgeMiddle;
-        }
-
-        threeStyleModel
-            .findAll(query)
-            .then((result) => {
-                const ans = {
-                    success: {
-                        code: 200,
-                        result,
-                    },
-                };
-
-                logger.emit('api.request', {
-                    requestType: 'GET',
-                    endpoint: '/hinemos/threeStyle/',
-                    params: {
-                        userName,
-                        part,
-                        buffer,
-                        sticker1,
-                        sticker2,
-                    },
-                    status: 'success',
-                    code: 200,
-                    msg: '',
-                });
-
-                res.json(ans);
-                res.status(200);
-            })
-            .catch((err) => {
-                logger.emit('api.request', {
-                    requestType: 'GET',
-                    endpoint: '/hinemos/threeStyle/',
-                    params: {
-                        userName,
-                        part,
-                        buffer,
-                        sticker1,
-                        sticker2,
-                    },
-                    status: 'error',
-                    code: 400,
-                    msg: err,
-                });
-
-                res.status(400).send(badRequestError);
-            });
-    });
+    app.get(`${process.env.EXPRESS_ROOT}/threeStyle/:part`, route.threeStyle.getProcess);
 
     // lettersから3-styleを引く
     app.get(process.env.EXPRESS_ROOT + '/threeStyleFromLetters/:part', (req, res, next) => {
@@ -1418,118 +1318,7 @@ sequelize.sync().then(() => {
         });
     });
 
-    app.post(process.env.EXPRESS_ROOT + '/threeStyle/:part', (req, res, next) => {
-        const userName = req.decoded.userName;
-        const part = req.params.part;
-
-        const buffer = req.body.buffer.replace(/\s*$/, '').replace(/^\s*/, '');
-        const sticker1 = req.body.sticker1.replace(/\s*$/, '').replace(/^\s*/, '');
-        const sticker2 = req.body.sticker2.replace(/\s*$/, '').replace(/^\s*/, '');
-        const setup = req.body.setup.replace(/\s*$/, '').replace(/^\s*/, '');
-        const move1 = req.body.move1.replace(/\s*$/, '').replace(/^\s*/, '');
-        const move2 = req.body.move2.replace(/\s*$/, '').replace(/^\s*/, '');
-
-        const okCond1 = (move1 !== '' && move2 !== '');
-        const okCond2 = (move1 === '' && move2 === '' && setup !== '');
-
-        if (!userName || !buffer || !sticker1 || !sticker2 || !(okCond1 || okCond2) || !(part === 'corner' || part === 'edgeMiddle')) {
-            logger.emit('api.request', {
-                requestType: 'POST',
-                endpoint: '/hinemos/threeStyle/',
-                params: {
-                    part,
-                    buffer,
-                    sticker1,
-                    sticker2,
-                    setup,
-                    move1,
-                    move2,
-                    userName: req.decoded.userName,
-                    decoded: req.decoded,
-                },
-                status: 'error',
-                code: 400,
-                msg: '',
-            });
-
-            res.status(400).send(badRequestError);
-            return;
-        }
-
-        const numberOfMoves = utils.getNumberOfMoves(setup, move1, move2);
-        const stickers = `${buffer} ${sticker1} ${sticker2}`;
-
-        let threeStyleModel;
-        if (part === 'corner') {
-            threeStyleModel = ThreeStyleCorner;
-        } else if (part === 'edgeMiddle') {
-            threeStyleModel = ThreeStyleEdgeMiddle;
-        }
-
-        threeStyleModel
-            .create({
-                userName,
-                numberOfMoves,
-                buffer,
-                sticker1,
-                sticker2,
-                stickers,
-                setup,
-                move1,
-                move2,
-            })
-            .then((result) => {
-                const ans = {
-                    success: {
-                        code: 200,
-                        result,
-                    },
-                };
-
-                logger.emit('api.request', {
-                    requestType: 'POST',
-                    endpoint: '/hinemos/threeStyle/',
-                    params: {
-                        part,
-                        buffer,
-                        sticker1,
-                        sticker2,
-                        setup,
-                        move1,
-                        move2,
-                        userName: req.decoded.userName,
-                        decoded: req.decoded,
-                    },
-                    status: 'success',
-                    code: 200,
-                    msg: '',
-                });
-                res.json(ans);
-                res.status(200);
-            })
-            .catch((err) => {
-                logger.emit('api.request', {
-                    requestType: 'POST',
-                    endpoint: '/hinemos/threeStyle/',
-                    params: {
-                        part,
-                        buffer,
-                        sticker1,
-                        sticker2,
-                        setup,
-                        move1,
-                        move2,
-                        userName: req.decoded.userName,
-                        decoded: req.decoded,
-                    },
-                    status: 'error',
-                    code: 400,
-                    msg: err,
-                });
-
-                res.status(400).send(badRequestError);
-            });
-    });
+    app.post(`${process.env.EXPRESS_ROOT}/threeStyle/:part`, route.threeStyle.postProcess);
 
     // 部分更新することは考えていないので、そのユーザのデータを全置き換え
     app.post(process.env.EXPRESS_ROOT + '/numbering/:part/', (req, res, next) => {
